Move ServiceTaskQueryCloudRequestModel to a newe file
