+++ lib/process-services-cloud/src/lib/task/task-list/models/filter-cloud-model.ts
@@ -80,33 +80,3 @@ export class TaskQueryCloudRequestModel {
         }
     }
 }
-
-export interface ServiceTaskQueryCloudRequestModel {
-    appName: string;
-    appVersion?: string;
-    id?: string;
-    status?: string;
-    maxItems: number;
-    skipCount: number;
-    sorting?: TaskListCloudSortingModel[];
-    activityName?: string;
-    activityType?: string;
-    completedDate?: Date;
-    elementId?: string;
-    executionId?: string;
-    processDefinitionId?: string;
-    processDefinitionKey?: string;
-    processDefinitionVersion?: number;
-    processInstanceId?: string;
-    serviceFullName?: string;
-    serviceName?: string;
-    serviceVersion?: string;
-    startedDate?: Date;
-}
-
-export interface ServiceTaskIntegrationContextCloudModel extends ServiceTaskQueryCloudRequestModel {
-    errorDate?: Date;
-    errorClassName?: string;
-    errorCode?: string;
-    errorMessage?: string;
-}
