+++ lib/process-services-cloud/src/lib/task/task-list/public-api.ts
@@ -19,6 +19,7 @@ export * from './components/task-list-cloud.component';
 export * from './components/service-task-list-cloud.component';
 
 export * from './models/filter-cloud-model';
+export * from './models/service-task-cloud.model';
 export * from './models/task-list-sorting.model';
 export * from './models/task-preset-cloud.model';
